updated battle tracker to be component, cleaned up
